requirements.txt
Django
python-dotenv
djangorestframework-simplejwt
django-cors-headers==4.7.0